% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdalassoCV.R
\name{AdalassoCV}
\alias{AdalassoCV}
\title{Adaptive Lasso with automatic \eqn{\gamma} selection with Cross-Validation}
\usage{
AdalassoCV(X, y, gammaSeq = c(0.5, 1, 2, 5), standardize = FALSE,
  lambdaOption = "min", initEst = "OLS")
}
\arguments{
\item{X}{The matrix of regressors.}

\item{y}{The response variable.}

\item{gammaSeq}{A sequence of gamma values.}

\item{standardize}{Should the regressors be standardized?}

\item{lambdaOption}{Either the minimizer (min) or the lambda which
achieves the highest lambda with error under the 1-standard deviation
of the minimum error.}

\item{initEst}{Initial estimator for the adaptive Lasso. Either OLS
or ridge regression.}
}
\value{
The tuple (lambda, gamma) of minimizers of the cross-validated 
error curve for the adaptive Lasso with the estimated model.
}
\description{
Compute the best lambda and gamma sequences for the adaptive Lasso with 
Cross-Validation.
}
\keyword{LassoGroupProject}
