% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_data.R
\docType{data}
\name{hp_data}
\alias{hp_data}
\title{House price dataset}
\format{A data frame with 338 rows and 255 variables:
\describe{
item{MSSubClass}{The building class}
item{MSZoning}{The general zoning classification}
item{LotFrontage}{Linear feet of street connected to property}
item{LotArea}{Lot size in square feet}
item{Street}{Type of road access}
item{Alley}{Type of alley access}
item{LotShape}{General shape of property}
item{LandContour}{Flatness of the property}
item{Utilities}{Type of utilities available}
item{SalePrice}{Log of the original sale price}
item{LotConfig}{Lot configuration}
item{LandSlope}{Slope of property}
item{Neighborhood}{Physical locations within Ames city limits}
item{Condition1}{Proximity to main road or railroad}
item{Condition2}{Proximity to main road or railroad (if a second is present)}
item{BldgType}{Type of dwelling}
item{HouseStyle}{Style of dwelling}
item{OverallQual}{Overall material and finish quality}
item{OverallCond}{Overall condition rating}
item{YearBuilt}{Original construction date}
item{YearRemodAdd}{Remodel date}
item{RoofStyle}{Type of roof}
item{RoofMatl}{Roof material}
item{Exterior1st}{Exterior covering on house}
item{Exterior2nd}{Exterior covering on house (if more than one material)}
item{MasVnrType}{Masonry veneer type}
item{MasVnrArea}{Masonry veneer area in square feet}
item{ExterQual}{Exterior material quality}
item{ExterCond}{Present condition of the material on the exterior}
item{Foundation}{Type of foundation}
item{BsmtQual}{Height of the basement}
item{BsmtCond}{General condition of the basement}
item{BsmtExposure}{Walkout or garden level basement walls}
item{BsmtFinType1}{Quality of basement finished area}
item{BsmtFinSF1}{Type 1 finished square feet}
item{BsmtFinType2}{Quality of second finished area (if present)}
item{BsmtFinSF2}{Type 2 finished square feet}
item{BsmtUnfSF}{Unfinished square feet of basement area}
item{TotalBsmtSF}{Total square feet of basement area}
item{Heating}{Type of heating}
item{HeatingQC}{Heating quality and condition}
item{CentralAir}{Central air conditioning}
item{Electrical}{Electrical system}
item{1stFlrSF}{First Floor square feet}
item{2ndFlrSF}{Second floor square feet}
item{LowQualFinSF}{Low quality finished square feet (all floors)}
item{GrLivArea}{Above grade (ground) living area square feet}
item{BsmtFullBath}{Basement full bathrooms}
item{BsmtHalfBath}{Basement half bathrooms}
item{FullBath}{Full bathrooms above grade}
item{HalfBath}{Half baths above grade}
item{Bedroom}{Number of bedrooms above basement level}
item{Kitchen}{Number of kitchens}
item{KitchenQual}{Kitchen quality}
item{TotRmsAbvGrd}{Total rooms above grade (does not include bathrooms)}
item{Functional}{Home functionality rating}
item{Fireplaces}{Number of fireplaces}
item{FireplaceQu}{Fireplace quality}
item{GarageType}{Garage location}
item{GarageYrBlt}{Year garage was built}
item{GarageFinish}{Interior finish of the garage}
item{GarageCars}{Size of garage in car capacity}
item{GarageArea}{Size of garage in square feet}
item{GarageQual}{Garage quality}
item{GarageCond}{Garage condition}
item{PavedDrive}{Paved driveway}
item{WoodDeckSF}{Wood deck area in square feet}
item{OpenPorchSF}{Open porch area in square feet}
item{EnclosedPorch}{Enclosed porch area in square feet}
item{3SsnPorch}{Three season porch area in square feet}
item{ScreenPorch}{Screen porch area in square feet}
item{PoolArea}{Pool area in square feet}
item{PoolQC}{Pool quality}
item{Fence}{Fence quality}
item{MiscFeature}{Miscellaneous feature not covered in other categories}
item{MiscVal}{Value of miscellaneous feature}
item{SaleType}{Type of sale}
item{SaleCondition}{Condition of sale}
}}
\source{
\url{De Cock D. Ames, Iowa: 
Alternative to the Boston housing 
data as an end of semester 
regression project[J]. Journal of 
Statistics Education, 2011, 19(3).}
}
\description{
A dataset containing information about the residential home market in
Ames, Iowa, United States in 2009.
}
