% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelSim.R
\name{ModelSelSim}
\alias{ModelSelSim}
\title{Replicate the results of the model selection consistency study}
\usage{
ModelSelSim(path = getwd(), scenario = c("big", "small"), nObs = 1000,
  p = 50, nSim = 100, nDesigns = 100, sigma = 0.1, adaGamma = 2,
  seed = 481)
}
\arguments{
\item{path}{The path to which you want the graphics to be exported to.}

\item{scenario}{Big or small coefficients scenarios to be simulated. Default are both.}

\item{nObs}{Number of observations.}

\item{p}{Number of variables.}

\item{nSim}{Number of simulations.}

\item{nDesigns}{Number of design matrices.}

\item{sigma}{Standard deviation of the noise for the regression model.}

\item{adaGamma}{Gamma parameter for the adaptive Lasso. Default is 2.}

\item{seed}{Seed for the simulations}
}
\value{
A (Two) pdf-file(s) exported to your path
}
\description{
Replicate the results of the Monte Carlo Simulation for the model selection
consistency study. The default values are set such that the results you obtain
are the same as in our term paper.
}
\keyword{LassoGroupProject}
