% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredContest.R
\name{PredContest}
\alias{PredContest}
\title{Replicate the results of the prediction contest simulation study}
\usage{
PredContest(path = getwd(), pList = c(50, 100), sigmaList = c(1, 5, 25),
  nObs = 500, nSim = 100, adaList = c(0.01, 0.1, 0.5, 1, 2, 5, 7),
  seed = 823)
}
\arguments{
\item{path}{The path to which you want the results to be exported to.}

\item{pList}{Vector with the choices of number of variables.}

\item{sigmaList}{Vector with the choices of noise levels.}

\item{nObs}{Number of observations.}

\item{nSim}{Number of simulations.}

\item{adaList}{Sequence of gamma parameters for the 
adaptive Lasso to cross-validate.}

\item{seed}{Seed for the simulations.}
}
\value{
(Number of variable choices x Number of noise levels) plots and 
tables in LaTeX and html. Moreover, the results are also saved in a list
if you assign a variable to the function as in the example below. The 
results are also exported to the directory specified in the 'path' argument.
}
\description{
Replicate the results of the Monte Carlo Simulation for the prediction
contest. The default values are set such that the results you obtain
are the same as in our term paper.
}
\details{
This function creates a new directory in the directory specified
in the 'path' argument named 'Prediction_Contest'. This directory in turn
include two subdirectories, 'Plots' and 'Tables'. Finally, in the 'Tables'
directory are included two directories more, 'HTML' for the result tables
in html-format and 'LaTeX' for the result tables in LaTeX format.
}
\examples{
results <- PredContest() # It takes long time with default arguments (around 5 hours)
}
\keyword{LassoGroupProject}
