% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP_Application.R
\name{SP_Application}
\alias{SP_Application}
\title{Replicate the results of the stock prices application}
\usage{
SP_Application(path = getwd(), adaList = c(0.01, 0.1, 0.5, 1, 2, 3),
  mccv = TRUE, cvIter = 100, seed = 123)
}
\arguments{
\item{path}{The path to which you want the results to be exported to.}

\item{adaList}{Sequence of gamma parameters for the 
adaptive Lasso to cross-validate.}

\item{mccv}{Should Monte Carlo Cross Validation 
be applied for evaluation?}

\item{cvIter}{Number of iterations for the Monte Carlo 
Cross Validation.}

\item{seed}{Seed for the simulations.}
}
\value{
Result tables and plots.
}
\description{
Replicate the results of the real data application with the 
stock price dataset.
}
\details{
This function creates a new directory, 'Application_SP' where
the results will be saved.
}
\examples{
results <- SP_Application()
}
\keyword{LassoGroupProject}
