% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_data.R
\docType{data}
\name{sp_data}
\alias{sp_data}
\title{S&P dataset}
\format{A data frame with 231 rows and 76 variables:
\describe{
 \item{After Tax ROE}{Return on Equity = Net 
 Income/Shareholder's Equity}
}}
\source{
\url{https://www.kaggle.com/camnugent/sandp500/data}
}
\description{
A dataset containing information about stock prices
and relevant information from the financial statements
of large companies which have common stocks listed
on NYSE and NASDAQ.
}
