% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdalassoCV.R
\name{AdalassoCV}
\alias{AdalassoCV}
\title{Adaptive Lasso with automatic \eqn{\gamma} selection by cross-validation}
\usage{
AdalassoCV(X, y, gammaSeq = c(0.5, 1, 2, 5), standardize = FALSE,
  lambdaOption = "min", initEst = "OLS")
}
\arguments{
\item{X}{A matrix of regressors.}

\item{y}{The response variable.}

\item{gammaSeq}{A sequence of gamma values to cross-validate.}

\item{standardize}{Should the regressors be standardized?}

\item{lambdaOption}{Should the lambda be the minimizer of the cross-validated
error curve (min) or should the lambda be chosen such that its corresponding
cross-validated error is only 1-standard-deviation above the minimum error (1se)?}

\item{initEst}{Initial estimator for the adaptive Lasso. 
Either OLS (OLS) or ridge regression (ridge).}
}
\value{
The tuple (\eqn{\lambda}, \eqn{\gamma}) of minimizers of the cross-validated 
error curve for the adaptive Lasso together with the estimated model.
}
\description{
Computes the best (\eqn{\lambda}, \eqn{\gamma}) pair for the adaptive Lasso by 
cross-validation.
}
\keyword{LassoGroupProject}
