% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVar_fun.R
\name{AVar_fun}
\alias{AVar_fun}
\title{Compute the Average Variance of an estimator}
\usage{
AVar_fun(modelFit)
}
\arguments{
\item{modelFit}{A matrix of fitting values of dimensions (number of observations x 
number of simulations)}
}
\value{
A numeric value
}
\description{
Computes the Average Variance (AV) of an estimator
in a Monte Carlo simulations study.
}
\keyword{LassoGroupProject}
