% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP_Application.R
\name{SP_Application}
\alias{SP_Application}
\title{Replicate the results of the stock prices application}
\usage{
SP_Application(path = getwd(), adaGamma = 1, mccv = TRUE, cvIter = 100,
  seed = 123)
}
\arguments{
\item{path}{The path to which you want the results to be exported to.}

\item{adaGamma}{Gamma parameter for the adaptive Lasso. Default is 1.}

\item{mccv}{Should Monte Carlo Cross Validation be applied for evaluation? Default is TRUE.}

\item{cvIter}{Number of iterations for the Monte Carlo Cross Validation. Default is 100.}

\item{seed}{Seed for the simulations.}
}
\value{
Result tables and plots.
}
\description{
Replicate the results of the real data application with the 
S&P data set.
}
\details{
This function creates a new directory, 'Application_SP' where
the results will be saved.
}
\keyword{LassoGroupProject}
